Conflicts:
	RemoteSpiNNaker/RemoteSpiNNakerJobProcessManager/src/main/java/uk/ac/manchester/cs/spinnaker/jobprocess/PyNNJobProcess.java
	RemoteSpiNNaker/RemoteSpiNNakerModel/src/main/java/uk/ac/manchester/cs/spinnaker/job/pynn/PyNNJobParameters.java

+++ RemoteSpiNNaker/RemoteSpiNNakerWeb/src/main/java/uk/ac/manchester/cs/spinnaker/jobmanager/XenVMExecuterFactory.java
@@ -28,39 +28,99 @@ import com.xensource.xenapi.VBD;
 import com.xensource.xenapi.VDI;
 import com.xensource.xenapi.VM;
 
+/**
+ * Executer factory that uses Xen VMs.
+ */
 public class XenVMExecuterFactory implements JobExecuterFactory {
+
     /** Bytes in a gigabyte. Well, a gibibyte, but that's a nasty word. */
     private static final long GB = 1024L * 1024L * 1024L;
     /** Time (in ms) between checks to see if a VM is running. */
     private static final int VM_POLL_INTERVAL = 10000;
 
+    /**
+     * Lock object used for synchronisation.
+     */
     private final Object lock = new Object();
+
+    /**
+     * The thread group of any threads.
+     */
     private final ThreadGroup threadGroup;
+
+    /**
+     * Logging.
+     */
     private final Logger logger = getLogger(getClass());
 
+    /**
+     * The URL of the Xen Server.
+     */
     @Value("${xen.server.url}")
     private URL xenServerUrl;
+
+    /**
+     * The username under which to control Xen.
+     */
     @Value("${xen.server.username}")
     private String username;
+
+    /**
+     * The password for the username.
+     */
     @Value("${xen.server.password}")
     private String password;
+
+    /**
+     * The name of the Xen Template VM.
+     */
     @Value("${xen.server.templateVm}")
     private String templateLabel;
+
+    /**
+     * True if the VMs should be deleted on shutdown of the VM.
+     */
     @Value("${deleteJobsOnExit}")
     private boolean deleteOnExit;
+
+    /**
+     * True if the Xen VM should shutdown on exit of the job.
+     */
     @Value("${xen.server.shutdownOnExit}")
     private boolean shutdownOnExit;
+
+    /**
+     * True if the log of the job should upload as it is output.
+     */
     @Value("${liveUploadOutput}")
     private boolean liveUploadOutput;
+
+    /**
+     * True if a spinnaker machine should be requested.
+     */
     @Value("${requestSpiNNakerMachine}")
     private boolean requestSpiNNakerMachine;
+
+    /**
+     * The default size of the disk to attach to the Xen VM.
+     */
     @Value("${xen.server.diskspaceInGbs}")
     private long defaultDiskSizeInGbs;
+
+    /**
+     * The maximum number of VMs to create.
+     */
     @Value("${xen.server.maxVms}")
     private int maxNVirtualMachines;
 
+    /**
+     * The current number of VMs.
+     */
     private int nVirtualMachines = 0;
 
+    /**
+     * Create a new Xen Executor Factory.
+     */
     public XenVMExecuterFactory() {
         this.threadGroup = new ThreadGroup("XenVM");
     }
@@ -79,6 +139,9 @@ public class XenVMExecuterFactory implements JobExecuterFactory {
         }
     }
 
+    /**
+     * Wait for the VM to come up.
+     */
     private void waitToClaimVM() {
         synchronized (lock) {
             logger.debug(nVirtualMachines + " of " + maxNVirtualMachines
@@ -97,6 +160,7 @@ public class XenVMExecuterFactory implements JobExecuterFactory {
         }
     }
 
+    /** Callback when the executor is finished. */
     protected void executorFinished() {
         synchronized (lock) {
             nVirtualMachines--;
@@ -108,15 +172,43 @@ public class XenVMExecuterFactory implements JobExecuterFactory {
 
     /** Taming the Xen API a bit. */
     class XenConnection implements AutoCloseable {
+
+        /**
+         * The Xen connection.
+         */
         private Connection conn;
+
+        /**
+         * The ID of the executor on the VM.
+         */
         private final String id;
 
-        XenConnection(final String id) throws XenAPIException, XmlRpcException {
-            this.id = id;
+        /**
+         * Make a connection to the Xen server.
+         *
+         * @param idParam
+         *            The ID of the connection.
+         * @throws XenAPIException
+         *             something went wrong
+         * @throws XmlRpcException
+         *             something went wrong
+         */
+        XenConnection(final String idParam)
+                throws XenAPIException, XmlRpcException {
+            this.id = idParam;
             conn = new Connection(xenServerUrl);
             loginWithPassword(conn, username, password);
         }
 
+        /**
+         * Get a virtual machine.
+         *
+         * @return a VM handle
+         * @throws XmlRpcException
+         *             something went wrong
+         * @throws IOException
+         *             something went wrong
+         */
         VM getVirtualMachine() throws XmlRpcException, IOException {
             final Set<VM> vmsWithLabel = VM.getByNameLabel(conn, templateLabel);
             if (vmsWithLabel.isEmpty()) {
@@ -127,6 +219,17 @@ public class XenVMExecuterFactory implements JobExecuterFactory {
             return template.createClone(conn, templateLabel + "_" + id);
         }
 
+        /**
+         * Get the first block device attached to a VM.
+         *
+         * @param vm
+         *            the VM that the block device is attached to.
+         * @return the block device
+         * @throws XmlRpcException
+         *             something went wrong
+         * @throws IOException
+         *             something went wrong
+         */
         VBD getVirtualBlockDevice(final VM vm)
                 throws XmlRpcException, IOException {
             final Set<VBD> disks = vm.getVBDs(conn);
@@ -136,17 +239,52 @@ public class XenVMExecuterFactory implements JobExecuterFactory {
             return disks.iterator().next();
         }
 
+        /**
+         * Get the label for a disk image.
+         *
+         * @param vdi
+         *            The disk image
+         * @param suffix
+         *            The operational suffix
+         * @return The label
+         * @throws XmlRpcException
+         *             something went wrong
+         * @throws XenAPIException
+         *             something went wrong
+         */
         String getLabel(final VDI vdi, final String suffix)
                 throws XmlRpcException, XenAPIException {
             return vdi.getNameLabel(conn) + "_" + id + "_" + suffix;
         }
 
+        /**
+         * Get a base disk image.
+         *
+         * @param disk
+         *            The block device hosting the image.
+         * @return the disk image.
+         * @throws XmlRpcException
+         *             something went wrong
+         * @throws XenAPIException
+         *             something went wrong
+         */
         VDI getBaseVDI(final VBD disk) throws XmlRpcException, XenAPIException {
             final VDI vdi = disk.getVDI(conn);
             vdi.setNameLabel(conn, getLabel(vdi, "base"));
             return vdi;
         }
 
+        /**
+         * Create a derived disk image.
+         *
+         * @param baseVDI
+         *            The base disk image.
+         * @return the disk image.
+         * @throws XenAPIException
+         *             something went wrong
+         * @throws XmlRpcException
+         *             something went wrong
+         */
         VDI createVDI(final VDI baseVDI)
                 throws XenAPIException, XmlRpcException {
             final VDI.Record descriptor = new VDI.Record();
@@ -158,6 +296,19 @@ public class XenVMExecuterFactory implements JobExecuterFactory {
             return VDI.create(conn, descriptor);
         }
 
+        /**
+         * Create a block device.
+         *
+         * @param vm
+         *            The VM to attach the device to.
+         * @param vdi
+         *            The disk image to initialise the device with.
+         * @return the block device
+         * @throws XenAPIException
+         *             something went wrong
+         * @throws XmlRpcException
+         *             something went wrong
+         */
         VBD createVBD(final VM vm, final VDI vdi)
                 throws XenAPIException, XmlRpcException {
             final VBD.Record descriptor = new VBD.Record();
@@ -170,27 +321,92 @@ public class XenVMExecuterFactory implements JobExecuterFactory {
             return VBD.create(conn, descriptor);
         }
 
+        /**
+         * Adds boot data to a VM.
+         *
+         * @param vm
+         *            The VM to add the data to.
+         * @param key
+         *            The key for the data.
+         * @param value
+         *            The value for the data. Will be converted to string.
+         * @throws XenAPIException
+         *             something went wrong
+         * @throws XmlRpcException
+         *             something went wrong
+         */
         void addData(final VM vm, final String key, final Object value)
                 throws XenAPIException, XmlRpcException {
             vm.addToXenstoreData(conn, key, value.toString());
         }
 
+        /**
+         * Boot up a VM.
+         *
+         * @param vm
+         *            The VM to boot.
+         * @throws XenAPIException
+         *             something went wrong
+         * @throws XmlRpcException
+         *             something went wrong
+         */
         void start(final VM vm) throws XenAPIException, XmlRpcException {
             vm.start(conn, false, true);
         }
 
+        /**
+         * Destroy a VM. Shuts the VM down as well if required.
+         *
+         * @param vm
+         *            The VM to destroy.
+         * @throws XenAPIException
+         *             something went wrong
+         * @throws XmlRpcException
+         *             something went wrong
+         */
         void destroy(final VM vm) throws XenAPIException, XmlRpcException {
             vm.destroy(conn);
         }
 
-        void destroy(final VBD vm) throws XenAPIException, XmlRpcException {
-            vm.destroy(conn);
+        /**
+         * Destroy a block device.
+         *
+         * @param vbd
+         *            The block device to destroy.
+         * @throws XenAPIException
+         *             something went wrong
+         * @throws XmlRpcException
+         *             something went wrong
+         */
+        void destroy(final VBD vbd) throws XenAPIException, XmlRpcException {
+            vbd.destroy(conn);
         }
 
-        void destroy(final VDI vm) throws XenAPIException, XmlRpcException {
-            vm.destroy(conn);
+        /**
+         * Destroy a disk image.
+         *
+         * @param vdi
+         *            The disk image to destroy.
+         * @throws XenAPIException
+         *             something went wrong
+         * @throws XmlRpcException
+         *             something went wrong
+         */
+        void destroy(final VDI vdi) throws XenAPIException, XmlRpcException {
+            vdi.destroy(conn);
         }
 
+        /**
+         * Get the state of a VM.
+         *
+         * @param vm
+         *            The VM to get the state of.
+         * @return The state.
+         * @throws XenAPIException
+         *             something went wrong
+         * @throws XmlRpcException
+         *             something went wrong
+         */
         VmPowerState getState(final VM vm)
                 throws XenAPIException, XmlRpcException {
             return vm.getPowerState(conn);
@@ -209,43 +425,93 @@ public class XenVMExecuterFactory implements JobExecuterFactory {
         }
     }
 
+    /**
+     * The executer core connector.
+     */
     class Executer implements JobExecuter, Runnable {
         // Parameters from constructor
+        /**
+         * The Job Manager to report to.
+         */
         private final JobManager jobManager;
+
+        /**
+         * The UUID to assign to this executor.
+         */
         private final String uuid;
+
+        /**
+         * The URL to give to the job process manager for it to speak to the
+         * job manager.
+         */
         private final URL jobProcessManagerUrl;
+
+        /**
+         * The arguments for the job process manager.
+         */
         private final String args;
 
         // Internal entities
+        /**
+         * The VM of the job.
+         */
         private VM clonedVm;
+
+        /**
+         * The disk of the cloned VM.
+         */
         private VBD disk;
+
+        /**
+         * The Virtual Disk Image of the cloned VM.
+         */
         private VDI vdi;
+
+        /**
+         * The Virtual Disk Image added to the VM.
+         */
         private VDI extraVdi;
+
+        /**
+         * The disk added to the VM.
+         */
         private VBD extraDisk;
 
-        Executer(final JobManager jobManager, final URL baseUrl)
+        /**
+         * Instantiate a connector.
+         *
+         * @param jobManagerParam
+         *            The job manager.
+         * @param baseUrl
+         *            the service root URL.
+         * @throws XmlRpcException
+         *             something went wrong
+         * @throws IOException
+         *             something went wrong
+         */
+        Executer(final JobManager jobManagerParam, final URL baseUrl)
                 throws XmlRpcException, IOException {
-            this.jobManager = jobManager;
+            this.jobManager = jobManagerParam;
             uuid = UUID.randomUUID().toString();
             jobProcessManagerUrl =
                     new URL(baseUrl, "job/" + JOB_PROCESS_MANAGER_ZIP);
 
-            final StringBuilder args = new StringBuilder("-jar ");
-            args.append(JOB_PROCESS_MANAGER_JAR);
-            args.append(" --serverUrl ");
-            args.append(baseUrl);
-            args.append(" --executerId ");
-            args.append(uuid);
+            final StringBuilder execArgs = new StringBuilder("-jar ");
+            execArgs.append(JOB_PROCESS_MANAGER_JAR);
+            execArgs.append(" --serverUrl ");
+            execArgs.append(baseUrl);
+            execArgs.append(" --executerId ");
+            execArgs.append(uuid);
             if (deleteOnExit) {
-                args.append(" --deleteOnExit");
+                execArgs.append(" --deleteOnExit");
             }
             if (liveUploadOutput) {
-                args.append(" --liveUploadOutput");
+                execArgs.append(" --liveUploadOutput");
             }
             if (requestSpiNNakerMachine) {
-                args.append(" --requestMachine");
+                execArgs.append(" --requestMachine");
             }
-            this.args = args.toString();
+            this.args = execArgs.toString();
         }
 
         @Override
@@ -258,9 +524,18 @@ public class XenVMExecuterFactory implements JobExecuterFactory {
             new Thread(threadGroup, this, "Executer (" + uuid + ")").start();
         }
 
-        synchronized XenConnection createVm()
+        /**
+         * Make a VM and launch it.
+         *
+         * @param conn
+         *            The underlying connection to the Xen server.
+         * @throws XmlRpcException
+         *             something went wrong
+         * @throws IOException
+         *             something went wrong
+         */
+        synchronized void createVm(final XenConnection conn)
                 throws XmlRpcException, IOException {
-            final XenConnection conn = new XenConnection(uuid);
             clonedVm = conn.getVirtualMachine();
             disk = conn.getVirtualBlockDevice(clonedVm);
             vdi = conn.getBaseVDI(disk);
@@ -273,9 +548,15 @@ public class XenVMExecuterFactory implements JobExecuterFactory {
                 conn.addData(clonedVm, "vm-data/shutdown", true);
             }
             conn.start(clonedVm);
-            return conn;
         }
 
+        /**
+         * Delete a VM for a job.
+         *
+         * @param conn The connection to the Xen server to use.
+         * @throws XenAPIException If there is a Xen API issue
+         * @throws XmlRpcException If there is an error speaking to Xen
+         */
         private synchronized void deleteVm(final XenConnection conn)
                 throws XenAPIException, XmlRpcException {
             if (conn == null) {
@@ -300,7 +581,10 @@ public class XenVMExecuterFactory implements JobExecuterFactory {
 
         @Override
         public void run() {
-            try (XenConnection conn = createVm()) {
+            XenConnection conn = null;
+            try {
+                conn = new XenConnection(uuid);
+                createVm(conn);
                 try {
                     VmPowerState powerState;
                     do {
@@ -315,18 +599,17 @@ public class XenVMExecuterFactory implements JobExecuterFactory {
                 } finally {
                     jobManager.setExecutorExited(uuid, null);
                 }
-
+            } catch (final Exception e) {
+                logger.error("Error setting up VM", e);
+                jobManager.setExecutorExited(uuid, e.getMessage());
+            } finally {
                 try {
-                    if (deleteOnExit) {
+                    if (conn != null && deleteOnExit) {
                         deleteVm(conn);
                     }
                 } catch (final Exception e) {
                     logger.error("Error deleting VM");
                 }
-            } catch (final Exception e) {
-                logger.error("Error setting up VM", e);
-                jobManager.setExecutorExited(uuid, e.getMessage());
-            } finally {
                 executorFinished();
             }
         }
